+++ test/functional/vimscript/eval_spec.lua
@@ -16,7 +16,6 @@ local lfs = require('lfs')
 local clear = helpers.clear
 local eq = helpers.eq
 local exc_exec = helpers.exc_exec
-local exec = helpers.exec
 local exec_lua = helpers.exec_lua
 local exec_capture = helpers.exec_capture
 local eval = helpers.eval
@@ -150,79 +149,6 @@ describe('List support code', function()
       eq(nil, ('Took too long to cancel: %g >= %g'):format(t_dur, dur / 8))
     end
   end)
-end)
-
--- oldtest: Test_deep_nest()
-it('Error when if/for/while/try/function is nested too deep',function()
-  clear()
-  local screen = Screen.new(80, 24)
-  screen:attach()
-  meths.set_option('laststatus', 2)
-  exec([[
-    " Deep nesting of if ... endif
-    func Test1()
-      let @a = join(repeat(['if v:true'], 51), "\n")
-      let @a ..= "\n"
-      let @a ..= join(repeat(['endif'], 51), "\n")
-      @a
-      let @a = ''
-    endfunc
-
-    " Deep nesting of for ... endfor
-    func Test2()
-      let @a = join(repeat(['for i in [1]'], 51), "\n")
-      let @a ..= "\n"
-      let @a ..= join(repeat(['endfor'], 51), "\n")
-      @a
-      let @a = ''
-    endfunc
-
-    " Deep nesting of while ... endwhile
-    func Test3()
-      let @a = join(repeat(['while v:true'], 51), "\n")
-      let @a ..= "\n"
-      let @a ..= join(repeat(['endwhile'], 51), "\n")
-      @a
-      let @a = ''
-    endfunc
-
-    " Deep nesting of try ... endtry
-    func Test4()
-      let @a = join(repeat(['try'], 51), "\n")
-      let @a ..= "\necho v:true\n"
-      let @a ..= join(repeat(['endtry'], 51), "\n")
-      @a
-      let @a = ''
-    endfunc
-
-    " Deep nesting of function ... endfunction
-    func Test5()
-      let @a = join(repeat(['function X()'], 51), "\n")
-      let @a ..= "\necho v:true\n"
-      let @a ..= join(repeat(['endfunction'], 51), "\n")
-      @a
-      let @a = ''
-    endfunc
-  ]])
-  screen:expect({any = '%[No Name%]'})
-  feed(':call Test1()<CR>')
-  screen:expect({any = 'E579: '})
-  feed('<C-C>')
-  screen:expect({any = '%[No Name%]'})
-  feed(':call Test2()<CR>')
-  screen:expect({any = 'E585: '})
-  feed('<C-C>')
-  screen:expect({any = '%[No Name%]'})
-  feed(':call Test3()<CR>')
-  screen:expect({any = 'E585: '})
-  feed('<C-C>')
-  screen:expect({any = '%[No Name%]'})
-  feed(':call Test4()<CR>')
-  screen:expect({any = 'E601: '})
-  feed('<C-C>')
-  screen:expect({any = '%[No Name%]'})
-  feed(':call Test5()<CR>')
-  screen:expect({any = 'E1058: '})
 end)
 
 describe("uncaught exception", function()
